Crate sevenz_rust

Re-exports

pub use lzma_rust as lzma;

pub use nt_time;

Structs

Aes256Sha256Decoder

Aes256Sha256Encoder

AesEncoderOptions

Archive

BlockDecoder

Password

SeqReader

SevenZArchiveEntry

SevenZMethod

SevenZMethodConfiguration

SevenZReader

SevenZWriter

Writes a 7z file

SourceReader

StreamMap

SubStreamsInfo

Enums

Error

MethodOptions

Functions

compress

hepler function to compress src path to dest writer

compress_encypted

compress_to_path

hepler function to compress src path to dest path

compress_to_path_encrypted

decompress

decompress a source reader to [dest] path

decompress_file

decompress a 7z file

decompress_file_with_extract_fn

decompress_file_with_password

decompress a encrypted file with password

decompress_with_extract_fn

decompress_with_extract_fn_and_password

decompress_with_password

default_entry_extract_fn